feat: fixed "filling" components' height in chain element parametes popup
